Move all components into a single module. Move dialogue under assets.
