+++ src/assets/dialogue/parser.rs
@@ -1,9 +1,9 @@
-use crate::dialogue::Dialogue;
+use super::Dialogue;
 use amethyst::error::Error;
 use pest::Parser;
 
 #[derive(pest_derive::Parser)]
-#[grammar = "dialogue/dialogue.pest"]
+#[grammar = "assets/dialogue/dialogue.pest"]
 struct DialogueParser;
 
 pub fn parse(input: &str) -> Result<Dialogue, Error> {
@@ -12,14 +12,18 @@ pub fn parse(input: &str) -> Result<Dialogue, Error> {
 
     for token in root.next().unwrap().into_inner() {
         match token.as_rule() {
-            Rule::passage => ret.passages.push(reformat_passage(token.as_str())),
+            Rule::passage => {
+                let passage = reformat_passage(token.as_str());
+                debug_assert!(!passage.is_empty());
+                ret.passages.push(passage)
+            }
             Rule::EOI => (),
             _ => {
                 unreachable!();
             }
         }
     }
-
+    debug_assert!(!ret.passages.is_empty());
     Ok(ret)
 }
